requirements.txt
crewai==0.28.8 
crewai_tools==0.1.6 
langchain_community==0.0.29
openai>=1.0.0
typing>=3.7.4
python-dotenv>=0.19.0